requirements.txt
Django==4.1.7
djangorestframework==3.14.0
environs==9.5.0
sorl-thumbnail==12.9.0
Pillow==9.4.0
django-ckeditor==6.5.1
django-phonenumber-field==7.0.2
beautifulsoup4==4.11.2
requests==2.28.2
selenium==4.8.2
selenium-stealth==1.0.6
instaloader==4.9.6
whitenoise==6.4.0
